// src/app/modules/notification/notification.enum.ts
/* eslint-disable no-unused-vars */
export enum ENUM_NOTIFICATION_TYPE {
    TASK_CREATED = 'TASK_CREATED',
    TASK_ACCEPTED = 'TASK_ACCEPTED',
    TASK_COMPLETED = 'TASK_COMPLETED',
    EXTENSION_REQUEST = 'EXTENSION_REQUEST',
    GENERAL = 'general',
    BID_PLACED = 'BID_PLACED',
    OFFER_REJECTED = 'OFFER_REJECTED',
    TASK_OFFERED = 'TASK_OFFERED',
    BID_ACCEPTED = 'BID_ACCEPTED',
    CANCELLATION_REQUESTED = 'CANCELLATION_REQUESTED',
    CANCELLATION_REQUEST_APPROVED = 'CANCELLATION_REQUEST_APPROVED',
    CANCELLATION_REQUEST_REJECTED = 'CANCELLATION_REQUEST_REJECTED',
    EXTENSION_REQUEST_REJECTED = 'EXTENSION_REQUEST_REJECTED',
    EXTENSION_REQUEST_ACCEPTED = 'EXTENSION_REQUEST_ACCEPTED',
    SUPPORT_CREATED = 'SUPPORT_CREATED',
    TASK_CANCELLED = 'TASK_CANCELLED',
    TASK_STARTED = 'TASK_STARTED',
    BID_COUNTER = 'BID_COUNTER',
    BID_REJECTED = 'BID_REJECTED',
    REPORT_CREATED = 'REPORT_CREATED',
    REPORT_RESOLVED = 'REPORT_RESOLVED',
}
export const NOTIFICATION_ACTION = {
    VIEW: 'VIEW',
    LIST: 'LIST',
    ACTION: 'ACTION',
};

export const NOTIFICATION_ENTITY = {
    ORDER: 'ORDER',
    SHIFT: 'SHIFT',
    PAYMENT: 'PAYMENT',
    USER: 'USER',
    CAMPAIGN: 'CAMPAIGN',
    REVIEW: 'REVIEW',
    JOB: 'JOB',
    SUPPORT: 'SUPPORT',
    TASK: 'TASK',
    BID: 'BID',
    REPORT: 'REPORT',
} as const;

export type NOTIFICATION_ENTITY_TYPE =
    (typeof NOTIFICATION_ENTITY)[keyof typeof NOTIFICATION_ENTITY];